+++ tests/MeuValorLiquido.WebApp.Tests/GlobalUsings.cs
@@ -2,6 +2,7 @@ global using System.Net;
 global using System.Text;
 global using System.Text.RegularExpressions;
 global using FluentAssertions;
+global using MeuValorLiquido.Core.Money;
 global using MeuValorLiquido.Modules.Calculators;
 global using MeuValorLiquido.Modules.Calculators.Tax;
 global using MeuValorLiquido.WebApp.Infrastructure;